@@ -12,91 +12,122 @@
     "test:e2e": "playwright test",
     "test": "vitest --root src --run",
     "test:watch": "vitest --root src",
-    "lint:eslint": "yarn eslint --fix src",
+    "lint:eslint": "pnpm eslint --fix src",
     "lint:prettier": "prettier --write src",
-    "lint": "yarn lint:eslint && yarn lint:prettier",
+    "lint": "pnpm lint:eslint && pnpm lint:prettier",
     "tsc": "tsc",
     "commit": "cz",
     "prepare": "husky install",
     "postbuild": "next-sitemap --config next-sitemap.config.cjs"
   },
   "dependencies": {
-    "@commitlint/config-conventional": "^17.4.4",
-    "@headlessui/react": "^1.7.11",
-    "@heroicons/react": "^2.0.16",
-    "@hookform/resolvers": "^3.3.1",
-    "@radix-ui/react-context-menu": "^2.1.3",
-    "@radix-ui/react-dialog": "^1.0.2",
-    "@radix-ui/react-dropdown-menu": "^2.0.2",
-    "@radix-ui/react-hover-card": "^1.0.5",
-    "@radix-ui/react-label": "^2.0.2",
-    "@radix-ui/react-navigation-menu": "^1.1.1",
-    "@radix-ui/react-popover": "^1.0.5",
-    "@radix-ui/react-select": "^1.2.0",
-    "@radix-ui/react-slider": "^1.1.1",
-    "@radix-ui/react-slot": "^1.0.2",
-    "@radix-ui/react-switch": "^1.0.2",
-    "@supabase/auth-helpers-nextjs": "^0.8.1",
+    "@commitlint/config-conventional": "^17.8.1",
+    "@headlessui/react": "^1.7.19",
+    "@heroicons/react": "^2.1.5",
+    "@hookform/resolvers": "^3.9.0",
+    "@radix-ui/react-accordion": "^1.2.0",
+    "@radix-ui/react-alert-dialog": "^1.1.1",
+    "@radix-ui/react-aspect-ratio": "^1.1.0",
+    "@radix-ui/react-avatar": "^1.1.0",
+    "@radix-ui/react-checkbox": "^1.1.1",
+    "@radix-ui/react-collapsible": "^1.1.0",
+    "@radix-ui/react-context-menu": "^2.2.1",
+    "@radix-ui/react-dialog": "^1.1.1",
+    "@radix-ui/react-dropdown-menu": "^2.1.1",
+    "@radix-ui/react-hover-card": "^1.1.1",
+    "@radix-ui/react-label": "^2.1.0",
+    "@radix-ui/react-menubar": "^1.1.1",
+    "@radix-ui/react-navigation-menu": "^1.2.0",
+    "@radix-ui/react-popover": "^1.1.1",
+    "@radix-ui/react-progress": "^1.1.0",
+    "@radix-ui/react-radio-group": "^1.2.0",
+    "@radix-ui/react-scroll-area": "^1.1.0",
+    "@radix-ui/react-select": "^1.2.2",
+    "@radix-ui/react-separator": "^1.1.0",
+    "@radix-ui/react-slider": "^1.2.0",
+    "@radix-ui/react-slot": "^1.1.0",
+    "@radix-ui/react-switch": "^1.1.0",
+    "@radix-ui/react-tabs": "^1.1.0",
+    "@radix-ui/react-toast": "^1.2.1",
+    "@radix-ui/react-toggle": "^1.1.0",
+    "@radix-ui/react-toggle-group": "^1.1.0",
+    "@radix-ui/react-tooltip": "^1.1.2",
+    "@supabase/auth-helpers-nextjs": "^0.8.7",
     "@supabase/auth-helpers-react": "^0.3.1",
-    "@supabase/supabase-js": "^2.31.0",
-    "@tailwindcss/typography": "^0.5.9",
-    "@tanstack/react-query": "^4.24.10",
+    "@supabase/ssr": "^0.5.1",
+    "@supabase/supabase-js": "^2.45.3",
+    "@tailwindcss/forms": "^0.5.9",
+    "@tailwindcss/typography": "^0.5.15",
+    "@tanstack/react-query": "^4.36.1",
     "class-variance-authority": "^0.7.0",
-    "clsx": "^2.0.0",
+    "clsx": "^2.1.1",
+    "cmdk": "1.0.0",
+    "date-fns": "^3.6.0",
+    "embla-carousel-react": "^8.2.1",
+    "framer-motion": "^11.5.4",
+    "input-otp": "^1.2.4",
     "lucide-react": "^0.279.0",
-    "next": "^14.0.3",
+    "next": "^14.2.8",
+    "next-nprogress-bar": "^2.3.13",
+    "next-safe-action": "^7.9.0",
     "next-seo": "^5.15.0",
-    "next-sitemap": "^3.1.52",
+    "next-sitemap": "^3.1.55",
     "next-themes": "^0.2.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-hook-form": "^7.46.2",
+    "path-to-regexp": "^8.0.0",
+    "react": "^18.3.1",
+    "react-day-picker": "8.10.1",
+    "react-dom": "^18.3.1",
+    "react-hook-form": "^7.53.0",
     "react-hot-toast": "^2.4.1",
     "react-no-ssr": "^1.1.0",
+    "react-resizable-panels": "^2.1.2",
+    "recharts": "^2.12.7",
     "rooks": "^7.14.1",
-    "sonner": "^1.0.3",
+    "server-only": "^0.0.1",
+    "sonner": "^1.5.0",
     "tailwind-merge": "^1.14.0",
-    "tailwindcss": "^3.2.7",
+    "tailwindcss": "^3.4.10",
     "tailwindcss-animate": "^1.0.7",
     "url-join": "^5.0.0",
-    "zod": "^3.22.2"
+    "vaul": "^0.9.2",
+    "zod": "^3.23.8"
   },
   "devDependencies": {
-    "@commitlint/cli": "^17.4.4",
-    "@commitlint/cz-commitlint": "^17.4.4",
-    "@next/eslint-plugin-next": "^13.1.6",
-    "@playwright/test": "^1.31.0",
-    "@semantic-release/changelog": "^6.0.2",
+    "@commitlint/cli": "^17.8.1",
+    "@commitlint/cz-commitlint": "^17.8.1",
+    "@next/eslint-plugin-next": "^13.5.6",
+    "@playwright/test": "^1.47.0",
+    "@semantic-release/changelog": "^6.0.3",
     "@semantic-release/git": "^10.0.1",
-    "@testing-library/react": "^14.0.0",
-    "@types/jsonwebtoken": "^9.0.1",
-    "@types/nprogress": "^0.2.0",
-    "@types/react": "^18.2.38",
-    "@types/react-dom": "^18.2.17",
-    "@typescript-eslint/eslint-plugin": "^6.12.0",
-    "@typescript-eslint/parser": "^6.12.0",
+    "@testing-library/react": "^14.3.1",
+    "@types/jsonwebtoken": "^9.0.6",
+    "@types/nprogress": "^0.2.3",
+    "@types/react": "^18.3.5",
+    "@types/react-dom": "^18.3.0",
+    "@typescript-eslint/eslint-plugin": "^6.21.0",
+    "@typescript-eslint/parser": "^6.21.0",
     "@vitejs/plugin-react": "^3.1.0",
     "commitizen": "^4.3.0",
     "concurrently": "^7.6.0",
     "env-cmd": "^10.1.0",
-    "eslint": "^8.54.0",
+    "eslint": "^8.57.0",
     "eslint-config-esnext": "^4.1.0",
     "eslint-config-node": "^4.1.0",
-    "eslint-config-prettier": "^8.6.0",
-    "eslint-plugin-import": "^2.27.5",
+    "eslint-config-prettier": "^8.10.0",
+    "eslint-plugin-import": "^2.30.0",
     "eslint-plugin-playwright": "^0.12.0",
     "eslint-plugin-prettier": "^4.2.1",
     "husky": "^8.0.3",
-    "jsdom": "^21.1.0",
-    "lint-staged": "^13.1.2",
-    "npm-check-updates": "^16.7.5",
-    "postcss": "^8.4.21",
-    "prettier": "^2.8.4",
-    "semantic-release": "^20.1.0",
-    "supabase": "^1.38.6",
+    "jsdom": "^21.1.2",
+    "lint-staged": "^13.3.0",
+    "npm-check-updates": "^16.14.20",
+    "postcss": "^8.4.45",
+    "prettier": "^2.8.8",
+    "semantic-release": "^20.1.3",
+    "supabase": "^1.191.3",
     "tailwindcss-cli": "^0.1.2",
-    "tsc-files": "^1.1.3",
-    "typescript": "^5.3.2",
+    "tsc-files": "^1.1.4",
+    "typescript": "^5.5.4",
     "vitest": "^0.28.5"
   },
   "config": {
