@@ -1,5 +1,12 @@
 'use client';
-import { Button } from '@/components/ui/Button';
+
+import { Trash } from 'lucide-react';
+import { useAction } from 'next-safe-action/hooks';
+import { useRouter } from 'next/navigation';
+import { useRef, useState } from 'react';
+import { toast } from 'sonner';
+
+import { Button } from '@/components/ui/button';
 import {
   Dialog,
   DialogContent,
@@ -8,49 +15,41 @@ import {
   DialogHeader,
   DialogTitle,
   DialogTrigger,
-} from '@/components/ui/Dialog';
-import { useMutation } from '@tanstack/react-query';
-import { useRouter } from 'next/navigation';
-import { useRef, useState } from 'react';
-import { toast } from 'react-hot-toast';
-import Trash from 'lucide-react/dist/esm/icons/trash';
+} from '@/components/ui/dialog';
+import { deletePrivateItemAction } from '@/data/user/privateItems';
 
 type Props = {
   itemId: string;
-  deleteItemAction: (itemId: string) => Promise<void>;
 };
 
-export const ConfirmDeleteItemDialog = ({
-  itemId,
-  deleteItemAction,
-}: Props) => {
-  const [open, setOpen] = useState(false);
-  const toastRef = useRef<string | null>(null);
+export const ConfirmDeleteItemDialog = ({ itemId }: Props): JSX.Element => {
+  const [open, setOpen] = useState<boolean>(false);
+  const toastRef = useRef<string | number | undefined>(undefined);
   const router = useRouter();
-  const { mutate, isLoading } = useMutation(
-    async (id: string) => {
-      return deleteItemAction(id);
+
+  const { execute, status } = useAction(deletePrivateItemAction, {
+    onExecute: () => {
+      toastRef.current = toast.loading('Deleting item...');
     },
-    {
-      onMutate: () => {
-        const toastId = toast.loading('Deleting item');
-        toastRef.current = toastId;
-      },
-      onSuccess: () => {
-        toast.success('Item deleted', { id: toastRef.current });
-        toastRef.current = null;
-        router.refresh();
-        router.push('/');
-      },
-      onError: () => {
-        toast.error('Failed to delete item', { id: toastRef.current });
-        toastRef.current = null;
-      },
-      onSettled: () => {
-        setOpen(false);
-      },
-    }
-  );
+    onSuccess: () => {
+      toast.success('Item deleted', { id: toastRef.current });
+      toastRef.current = undefined;
+      router.refresh();
+      router.push('/');
+      setOpen(false);
+    },
+    onError: ({ error }) => {
+      const errorMessage =
+        error.serverError ?? error.fetchError ?? 'Failed to delete item';
+      toast.error(errorMessage, { id: toastRef.current });
+      toastRef.current = undefined;
+    },
+  });
+
+  const handleDelete = () => {
+    execute({ id: itemId });
+  };
+
   return (
     <Dialog open={open} onOpenChange={setOpen}>
       <DialogTrigger asChild>
@@ -69,20 +68,16 @@ export const ConfirmDeleteItemDialog = ({
           <Button
             type="button"
             variant="destructive"
-            disabled={isLoading}
-            onClick={() => {
-              mutate(itemId);
-            }}
+            disabled={status === 'executing'}
+            onClick={handleDelete}
           >
-            {isLoading ? `Deleting item...` : `Yes, delete`}
+            {status === 'executing' ? 'Deleting item...' : 'Yes, delete'}
           </Button>
           <Button
-            disabled={isLoading}
+            disabled={status === 'executing'}
             type="button"
             variant="outline"
-            onClick={() => {
-              setOpen(false);
-            }}
+            onClick={() => setOpen(false)}
           >
             Cancel
           </Button>
