@@ -1,54 +1,64 @@
 'use client';
 
-import { Button } from '@/components/ui/Button';
-import { insertItemAction } from '@/data/anon/items';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { useRef } from 'react';
 import { useRouter } from 'next/navigation';
-import { useRef, useState } from 'react';
-import { toast } from 'react-hot-toast';
+import { useAction } from 'next-safe-action/hooks';
+import { toast } from 'sonner';
+import { useForm } from 'react-hook-form';
+import { zodResolver } from '@hookform/resolvers/zod';
+import { z } from 'zod';
 
-export const ClientPage = () => {
+import { Button } from '@/components/ui/button';
+import { Input } from '@/components/ui/input';
+import { Textarea } from '@/components/ui/textarea';
+import { insertItemAction } from '@/data/anon/items';
+
+const formSchema = z.object({
+  name: z.string().min(1, 'Name is required'),
+  description: z.string().min(1, 'Description is required'),
+});
+
+type FormData = z.infer<typeof formSchema>;
+
+export const ClientPage: React.FC = () => {
   const router = useRouter();
-  const queryClient = useQueryClient();
-  const toastRef = useRef<string | null>(null);
+  const toastRef = useRef<string | number | undefined>(undefined);
 
-  const { mutate } = useMutation(
-    async (item: { name: string; description: string }) => {
-      return insertItemAction(item);
+  const {
+    register,
+    handleSubmit,
+    formState: { errors, isValid },
+  } = useForm<FormData>({
+    resolver: zodResolver(formSchema),
+    mode: 'onChange',
+  });
+
+  const { execute, status } = useAction(insertItemAction, {
+    onExecute: () => {
+      toastRef.current = toast.loading('Creating item');
     },
-    {
-      onMutate: () => {
-        const toastId = toast.loading('Creating item');
-        toastRef.current = toastId;
-      },
+    onSuccess: ({ data }) => {
+      toast.success('Item created', { id: toastRef.current });
+      toastRef.current = undefined;
+      router.refresh();
+      if (data) {
+        router.push(`/item/${data}`);
+      }
+    },
+    onError: ({ error }) => {
+      const errorMessage =
+        error.serverError ?? error.fetchError ?? 'Failed to create item';
+      toast.error(errorMessage, { id: toastRef.current });
+      toastRef.current = undefined;
+    },
+  });
 
-      onSuccess: (newItemId) => {
-        toast.success('Item created', { id: toastRef.current });
-        toastRef.current = null;
-        router.refresh();
-        queryClient.invalidateQueries(['items']);
-        router.push(`/item/${newItemId}`);
-      },
-      onError: () => {
-        toast.error('Failed to create item', { id: toastRef.current });
-        toastRef.current = null;
-      },
-    }
-  );
-  const [name, setName] = useState('');
-  const [description, setDescription] = useState('');
+  const onSubmit = (data: FormData) => {
+    execute(data);
+  };
+
   return (
-    <form
-      className="flex flex-col space-y-4"
-      onSubmit={(event) => {
-        event.preventDefault();
-        const formData = new FormData(event.target as HTMLFormElement);
-        //TODO: do better validation 🤷‍♂️
-        const name = formData.get('name') as string;
-        const description = formData.get('description') as string;
-        mutate({ name, description });
-      }}
-    >
+    <form onSubmit={handleSubmit(onSubmit)} className="flex flex-col space-y-4">
       <div>
         <h1 className="mt-1 text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl lg:text-5xl">
           Create Item
@@ -61,16 +71,15 @@ export const ClientPage = () => {
         >
           Name
         </label>
-        <input
-          value={name}
-          onChange={(event) => {
-            setName(event.target.value);
-          }}
+        <Input
+          {...register('name')}
           id="name"
-          name="name"
           type="text"
           className="block w-full appearance-none rounded-md border border-gray-300 px-3 py-2 placeholder-gray-400 shadow-sm focus:border-green-500 focus:outline-none focus:ring-green-500 sm:text-sm"
         />
+        {errors.name && (
+          <p className="text-red-500 text-sm mt-1">{errors.name.message}</p>
+        )}
       </div>
       <div className="space-y-2">
         <label
@@ -79,19 +88,24 @@ export const ClientPage = () => {
         >
           Description
         </label>
-        <textarea
-          value={description}
-          onChange={(event) => {
-            setDescription(event.target.value);
-          }}
+        <Textarea
+          {...register('description')}
           id="description"
-          name="description"
           rows={4}
           className="block w-full appearance-none rounded-md border border-gray-300 px-3 py-2 placeholder-gray-400 shadow-sm focus:border-green-500 focus:outline-none focus:ring-green-500 sm:text-sm"
         />
+        {errors.description && (
+          <p className="text-red-500 text-sm mt-1">
+            {errors.description.message}
+          </p>
+        )}
       </div>
-      <Button variant="default" type="submit">
-        Create Item
+      <Button
+        variant="default"
+        type="submit"
+        disabled={status === 'executing' || !isValid}
+      >
+        {status === 'executing' ? 'Creating Item...' : 'Create Item'}
       </Button>
     </form>
   );
