@@ -1,10 +1,7 @@
-import { cn } from '@/utils/cn';
+import { cn } from '@/lib/utils';
 import { ComponentProps } from 'react';
 
 export function Subtle({ className, ...rest }: ComponentProps<'p'>) {
-  const classNames = cn(
-    'text-sm text-slate-500 dark:text-slate-400',
-    className
-  );
+  const classNames = cn('text-sm text-muted-foreground', className);
   return <p className={classNames} {...rest}></p>;
 }
