@@ -1,11 +1,10 @@
 'use server';
-import { createSupabaseServerComponentClient } from '@/supabase-clients/createSupabaseServerComponentClient';
+import { createSupabaseClient } from '@/supabase-clients/server';
 import { Table } from '@/types';
-
 export const getAllPrivateItems = async (): Promise<
   Array<Table<'private_items'>>
 > => {
-  const supabase = createSupabaseServerComponentClient();
+  const supabase = createSupabaseClient();
   const { data, error } = await supabase.from('private_items').select('*');
 
   if (error) {
@@ -18,7 +17,7 @@ export const getAllPrivateItems = async (): Promise<
 export const getPrivateItem = async (
   id: string
 ): Promise<Table<'private_items'>> => {
-  const supabase = createSupabaseServerComponentClient();
+  const supabase = createSupabaseClient();
   const { data, error } = await supabase
     .from('private_items')
     .select('*')
